requirements.txt
Flask==3.0.3
tensorflow==2.18.0
Pillow==10.3.0
numpy==1.26.4
flask-cors==5.0.1